+++ magda-int-test/src/test/scala/au/csiro/data61/magda/indexer/WebhookSpecBase.scala
@@ -188,6 +188,7 @@ trait WebhookSpecBase extends BaseApiSpec with RegistryConverters with ModelProt
                       modifyJson(dist.toJson.asJsObject, Map(
                         "license" -> dist.license.flatMap(_.name).map(_.toJson).getOrElse(JsNull)))))).toJson),
             "source" -> dataSet.source.toJson,
+            "provenance" -> dataSet.provenance.toJson,
             "dataset-publisher" -> dataSet.publisher.map(publisher => JsObject(
               "publisher" -> JsObject(
                 "id" -> publisher.identifier.toJson,