+++ CHANGES.md
@@ -9,6 +9,8 @@ Registry:
 -   Changed database schema.
 -   Provide services based on Tenant IDs.
 -   Fixed PATCH request to registry won't trigger notification to webhook
+-   Moved provenance and information security information out of `dcat-dataset-strings`.
+-   Removed some unused fields from `dcat-dataset-strings` - it should now be back to looking more-or-less like DCAT.
 
 Gateway:
 
@@ -80,6 +82,7 @@ UI:
 -   Style adjustment for question Who can see the dataset once it is published on Access and Use page
 -   Added access location auto complete input on Access and Use page
 -   Fixed a JS error which causes blank screen on Organisations Page
+-   Fixed: drop a folder to Add dataset file drop area will break the UI
 -   Show an error message screen if the user is not allowed to access the add dataset page
 -   Make sure all publish new dataset errors are captured and shown to users on add dataset page